requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-telegram-bot==22.2
python-dotenv==1.0.0
pydantic==2.11.7
httpx==0.28.1
openai==1.58.1
anthropic==0.40.0
google-generativeai==0.8.3
supabase==2.9.1
aiofiles==24.1.0

